% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_df.R
\name{correct_df}
\alias{correct_df}
\title{Hardcode changes to web scraped PT cases and deaths}
\usage{
correct_df(
  data,
  metric = "",
  Jurisdiction = "",
  correction_date = "",
  corrected_value = ""
)
}
\arguments{
\item{data}{dataset you wish to correct. df_corrected data set as default}

\item{metric}{choice between updating "cases" or "deaths" = other text inputs will be ignored}

\item{correction_date}{date that you want to make correction for}

\item{corrected_value}{new input value}

\item{jurisdiction}{region that is being corrected. Only takes one input}
}
\value{

}
\description{
This function allows one to make corrections to webscraped case and death data with a more intuitive syntax.
}
\examples{

}
