% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_PT_names_functions.R
\name{factor_PT_west_to_east}
\alias{factor_PT_west_to_east}
\title{Factor PT names from West to East}
\usage{
factor_PT_west_to_east(
  input_data,
  geo_variable = "Jurisdiction",
  size = "small",
  Canada_first = FALSE
)
}
\arguments{
\item{input_data}{(required) an input data frame, tibble, etc.}

\item{geo_variable}{(optional) the variable that contains the information on jurisdiction. Note: must be a character. Defaults to "Jurisdiction"}

\item{size}{(optional) the format of the variable. "small" for abbreviations (ie. BC), "big" for complete words (ie. "British Columbia") defaults to "small".}

\item{Canada_first}{(optional) Boolean that determines whether Canada values should be placed first or last in the order. Defaults to last.}
}
\value{

}
\description{
This function converts a character "jurisdiction" variable to an ordered factor to help arrange PTs in figures and tables from west to east!
}
