% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table_functions.R
\name{format_casedeath_table}
\alias{format_casedeath_table}
\title{Format the Cases and Deaths table for presentation in the Daily Trend Report!}
\usage{
format_casedeath_table(input_table)
}
\arguments{
\item{input_table}{}
}
\value{

}
\description{
Format the Cases and Deaths table for presentation in the Daily Trend Report!
}
