% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table_functions.R
\name{format_labtesting_table}
\alias{format_labtesting_table}
\title{Format the lab testing table for presentation in the Daily Trend Report!}
\usage{
format_labtesting_table(input_table)
}
\arguments{
\item{input_table}{}
}
\value{

}
\description{
Format the lab testing table for presentation in the Daily Trend Report!
}
