% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_reports_functions.R
\name{generate_Nemer_report}
\alias{generate_Nemer_report}
\title{Generating the Nemer report}
\usage{
generate_Nemer_report(report_date = "")
}
\arguments{
\item{report_date}{}
}
\value{

}
\description{
Generating the Nemer report
}
