% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_reports_functions.R
\name{generate_trend_report}
\alias{generate_trend_report}
\title{Generating the daily trend report}
\usage{
generate_trend_report(report_date = "")
}
\arguments{
\item{report_date}{}
}
\value{

}
\description{
Generating the daily trend report
}
