% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_adjusted_case_death_data}
\alias{import_adjusted_case_death_data}
\title{Import corrected web-scraped case and death numbers, and apply hardcode corrections}
\usage{
import_adjusted_case_death_data()
}
\value{

}
\description{
This function returns web-scraped case and death data with additional corrections for data dumps, etc.
}
\details{
Note - cumulative totals will not line up with other sources as there are instances where PTs reported
cases and deaths without specifying a date. These data dumps get removed in this function, but we are not able to
reassign them accurately.
}
\examples{

df<-import_adjusted_case_death_data()


}
