% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_adjusted_infobase_data}
\alias{import_adjusted_infobase_data}
\title{Import corrected web-scraped case and death numbers from Infobase}
\usage{
import_adjusted_infobase_data()
}
\value{

}
\description{
This function downloads the covid19.csv dataset hosted at: \url{https://health-infobase.canada.ca/src/data/covidLive/covid19.csv}, but also makes additional corrections for data dumps, etc.
Important note - cumulative totals will not line up with other sources as there are instances where PTs reported
cases and deaths without specifying a date. These additional cases get removed in this function, but we are not able to
reassign them accurately.
}
\examples{

df<-import_adjusted_infobase_data()

}
\seealso{
\code{\link{import_raw_infobase_data}}

Other import functions: 
\code{\link{import_DISCOVER_data}()},
\code{\link{import_SALT_data}()},
\code{\link{import_demographic_data}()},
\code{\link{import_hosp_data}()},
\code{\link{import_international_data}()},
\code{\link{import_raw_infobase_data}()}
}
\concept{import functions}
