% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_case_death_data}
\alias{import_case_death_data}
\title{Import web-scraped case and death numbers, and apply hardcode corrections}
\usage{
import_case_death_data()
}
\value{

}
\description{
Import web-scraped case and death numbers, and apply hardcode corrections
}
