% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_case_report_form_data}
\alias{import_case_report_form_data}
\title{Import case report form data}
\usage{
import_case_report_form_data(method = "extract")
}
\arguments{
\item{method}{A string with options: 'extract' or 'metabaser' depending on the way you want to access the data}
}
\value{

}
\description{
Import case report form data
}
