% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_demographic_data}
\alias{import_demographic_data}
\title{get demographic data}
\usage{
import_demographic_data()
}
\value{
list with 2 values. First value: latest_can_pop. Second value: pt_pop_20
}
\description{
get demographic data
}
