% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_DISCOVER_data}
\alias{import_DISCOVER_data}
\title{Import case report form data from DISCOVER}
\usage{
import_DISCOVER_data(
  method = "extract",
  metabase_user = "",
  metabase_pass = ""
)
}
\arguments{
\item{method}{A string with options: 'extract' or 'metabaser' depending on the way you want to access the data}
}
\value{

}
\description{
Import case report form data from DISCOVER
}
\examples{

qry_cases_raw<-import_DISCOVER_data()

}
\seealso{
Other import functions: 
\code{\link{import_SALT_data}()},
\code{\link{import_adjusted_infobase_data}()},
\code{\link{import_demographic_data}()},
\code{\link{import_hosp_data}()},
\code{\link{import_international_data}()},
\code{\link{import_raw_infobase_data}()}
}
\concept{import functions}
