% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_international_data}
\alias{import_international_data}
\title{Import international data}
\usage{
import_international_data()
}
\value{

}
\description{
Import international data
}
\examples{

df_int<-import_international_data()

}
\seealso{
Other import functions: 
\code{\link{import_DISCOVER_data}()},
\code{\link{import_SALT_data}()},
\code{\link{import_adjusted_infobase_data}()},
\code{\link{import_demographic_data}()},
\code{\link{import_hosp_data}()},
\code{\link{import_raw_infobase_data}()}
}
\concept{import functions}
