% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_raw_case_death_data}
\alias{import_raw_case_death_data}
\title{Import raw web-scraped case and death numbers, and apply hardcode corrections}
\usage{
import_raw_case_death_data()
}
\value{

}
\description{
Import raw web-scraped case and death numbers, and apply hardcode corrections
}
\examples{
df_raw<-import_raw_case_death_data

}
