% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_raw_infobase_data}
\alias{import_raw_infobase_data}
\title{Import raw web-scraped case and death numbers from Infobase}
\usage{
import_raw_infobase_data()
}
\value{

}
\description{
This function downloads the covid19.csv dataset hosted at: https://health-infobase.canada.ca/src/data/covidLive/covid19.csv.
This dataset is maintained by the daily report epidemiologist from the Operations group at SED (former HPOC). Questions on contents of this dataset can be sent to hsfluepi generic account.
}
\examples{
df_raw<-import_raw_infobase_data()

}
\seealso{
\code{\link{import_adjusted_infobase_data}}

Other import functions: 
\code{\link{import_DISCOVER_data}()},
\code{\link{import_SALT_data}()},
\code{\link{import_adjusted_infobase_data}()},
\code{\link{import_demographic_data}()},
\code{\link{import_hosp_data}()},
\code{\link{import_international_data}()}
}
\concept{import functions}
