% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_SALT_data}
\alias{import_SALT_data}
\title{Import lab testing data from SALT}
\usage{
import_SALT_data()
}
\value{

}
\description{
Import lab testing data from SALT
}
\examples{
salt_raw<-import_SALT_data()

}
\seealso{
Other import functions: 
\code{\link{import_DISCOVER_data}()},
\code{\link{import_adjusted_infobase_data_backup}()},
\code{\link{import_adjusted_infobase_data}()},
\code{\link{import_demographic_data}()},
\code{\link{import_hosp_data}()},
\code{\link{import_international_data}()},
\code{\link{import_raw_infobase_data}()}
}
\concept{import functions}
