% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_PT_names_functions.R
\name{recode_PT_names_to_big}
\alias{recode_PT_names_to_big}
\title{recode_PT_names_to_big}
\usage{
recode_PT_names_to_big(dataset, geo_variable = "jurisdiction")
}
\arguments{
\item{geo_variable}{}
}
\value{

}
\description{
recode_PT_names_to_big
}
