% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_PT_names_functions.R
\name{recode_PT_names_to_small}
\alias{recode_PT_names_to_small}
\title{Recode_pt_names_to_small}
\usage{
recode_PT_names_to_small(dataset, geo_variable = "Jurisdiction")
}
\arguments{
\item{geo_variable}{}
}
\value{

}
\description{
Recode_pt_names_to_small
}
