% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{turn_char_vec_to_comma_list}
\alias{turn_char_vec_to_comma_list}
\title{Convert vector to readable list}
\usage{
turn_char_vec_to_comma_list(vector)
}
\arguments{
\item{vector}{(Required). The character vector of interest. Must be a character vector or this just won't work.}
}
\value{

}
\description{
This function takes as an input a character vector, and returns a human readable list, with commas and "ands". How nice!
}
