% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{turn_num_to_percent_change}
\alias{turn_num_to_percent_change}
\title{Title}
\usage{
turn_num_to_percent_change(input_dataset, numeric_variable, accuracy = 0.1)
}
\arguments{
\item{accuracy}{}
}
\value{

}
\description{
Title
}
