% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_adjusted_case_death_data}
\alias{import_adjusted_case_death_data}
\title{Import corrected web-scraped case and death numbers, and apply hardcode corrections}
\usage{
import_adjusted_case_death_data()
}
\value{

}
\description{
Corrections for data dumps, etc. Note - don't use cumulative totals from this, as they are missing some cases that could not be assigned specific dates.
}
