% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_demographic_data}
\alias{import_demographic_data}
\title{Get demographic data}
\usage{
import_demographic_data()
}
\value{
list with 2 values. First value: latest_can_pop. Second value: pt_pop_20.

Note - this function is optional, and is not currently recommended to use. These two datasets are also stored as R data in this package.
To access them, simply call: PHACTrendR::latest_can_pop, and PHACTrendR::pt_pop_20
}
\description{
Get demographic data
}
\examples{

demographic_data<-import_demographic_data()

latest_can_pop<-demographic_data[[1]]
pt_pop20<-demographic_data[[2]]

}
\seealso{
Other import functions: 
\code{\link{import_DISCOVER_data}()},
\code{\link{import_SALT_data}()},
\code{\link{import_adjusted_infobase_data_backup}()},
\code{\link{import_adjusted_infobase_data}()},
\code{\link{import_hosp_data}()},
\code{\link{import_international_data}()},
\code{\link{import_raw_infobase_data}()}
}
\concept{import functions}
