% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_functions.R
\name{import_hosp_data}
\alias{import_hosp_data}
\title{Import web scraped hospital and ICU data}
\usage{
import_hosp_data()
}
\value{

}
\description{
This function imports web-scraped data on hospitalization of Covid-19 patients in Canada. There is a correction done for AB.
}
\examples{

all_hosp_data<-import_hosp_data()


}
