% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_PT_names_functions.R
\name{recode_PT_names_to_small}
\alias{recode_PT_names_to_small}
\title{Recode_pt_names_to_small}
\usage{
recode_PT_names_to_small(dataset, geo_variable = "jurisdiction")
}
\arguments{
\item{dataset}{(required) the dataset of interest. Can be a data.frame, tibble, or even just a character vector}

\item{geo_variable}{(optional) if the dataset is not a character vector, designate the geographical variable of interest (defaults to "jurisdiction")}
}
\value{

}
\description{
Recode_pt_names_to_small
}
